deleting calfeed when deleting member

+++ lib/caldav.py
@@ -22,6 +22,8 @@ from django.utils.translation import gettext_lazy as _
 import os
 from django.core.files.storage import DefaultStorage
 from gig.util import GigStatusChoices
+from django.conf import settings
+
 
 filesys = FileSystemStorage("calfeeds", "calfeeds")
 
@@ -43,9 +45,10 @@ def get_calfeed(tag):
 
 
 def delete_calfeed(tag):
-    file_path = f'{tag}.txt'
-    if filesys.exists(file_path):
-        filesys.delete(file_path)
+    if settings.DYNAMIC_CALFEED:
+        file_path = f'{tag}.txt'
+        if filesys.exists(file_path):
+            filesys.delete(file_path)
 
 
 def make_calfeed(the_title, the_events, the_language, the_uid):
